+++ docs/README.md
@@ -25,7 +25,7 @@
     - [x] 이름은 중복을 허용하지 않는다.
     - [x] 이름은 공백을 허용하지 않는다.
     - [x] 자동차 이름은 빈 문자가 되어서는 안된다.
-- [ ] 참가자 입력이 성공적으로 완료되면 이동 횟수 입력 폼을 활성화 하고 참가자 입력폼을 비활성화한다..
+- [ ] 참가자 입력이 성공적으로 완료되면 이동 횟수 입력 폼을 활성화 하고 참가자 입력폼을 비활성화한다.
 
 ### 2. 이동 횟수 입력 기능
