+++ index.html
@@ -14,15 +14,15 @@
       <h1>🏎️ 자동차 경주 게임🏁</h1>
     </header>
     <section id="racing-form-container">
-      <form id="car-names-form">
-        <label>5자 이하의 이름을 콤마로 구분해서 입력해주세요.</label>
+      <form id="car-names-form" disabled>
+        <label for="car-names-input">5자 이하의 이름을 콤마로 구분해서 입력해주세요.</label>
         <input id="car-names-input" type="text" placeholder="east, west, south, north, all"/>
-        <button id="car-names-submit">확인</button>
+        <button id="car-names-submit" type = "submit">확인</button>
       </form>
       <form id="racing-count-form">
-        <label>시도할 횟수를 입력해주세요.</label>
-        <input id="racing-count-input"type="number" placeholder="10"/>
-        <button id="racing-count-submit">확인</button>
+        <label for="racing-count-input">시도할 횟수를 입력해주세요.</label>
+        <input id="racing-count-input"type="number" placeholder="10" disabled/>
+        <button id="racing-count-submit" type="submit" disabled>확인</button>
       </form>
     </section>
   </div>
